Add files via upload
Rock-Paper-Scissors Game

This interactive Rock-Paper-Scissors game is built using HTML, CSS, and JavaScript. 

HTML provides the structure of the game, including buttons for player choices and areas to display results. 

CSS enhances the visual appeal, offering a stylish and user-friendly interface with animations and responsive design to ensure a great experience on both desktop and mobile devices.

JavaScript powers the game logic, handling user inputs, generating random computer choices, and determining the winner. It dynamically updates the score and displays messages based on the outcome of each round.

Play against the computer, and may the best choice win!